pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-databases-agent"
version = "0.1.0"
description = "MCP server and CLI client with integrated tools and resources"
readme = "README.md"
license = {text = "MIT"}
authors = []
requires-python = ">=3.10"
dependencies = [
    "anthropic==0.77.0",
    "mcp==1.26.0",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/mcp-datasources"
Repository = "https://github.com/yourusername/mcp-datasources.git"

[tool.setuptools]
packages = ["mcp"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
